[libcudacxx][fpemu] fix MSVC C4805 bool/uint32 mix in __sar_64_rnd
MSVC /W4 /WX rejected `__res.x[0] |= __sticky` (uint32_t |= bool) as an
error, which broke every fpemu unit build across all MSVC configs. Cast
the sticky bool to uint32_t explicitly.

+++ libcudacxx/include/cuda/__fp/fpemu_impl.h
@@ -975,7 +975,7 @@ _CCCL_TRIVIAL_API __uint32x2 __sar_64_rnd(__uint32x2 __man, int __shift, bool __
     bool __sticky   = (__man64 & __mask) != 0;
     if constexpr (_Rm == __fpemu_rounding::rn)
     {
-      __res.x[0] |= __sticky;
+      __res.x[0] |= static_cast<uint32_t>(__sticky);
     }
     else if constexpr (_Rm == __fpemu_rounding::rz)
     {
